Changed all usages of ArrayAdapter to NoFilterArrayAdapter to persist AutoCompleteTextView suggestions after rotation
